V1 of the Poker Game (No GUI)
Does allow the user to select how many players there are, and how allows the user to see everyone's cards (as well as Community Cards). The winning logic is still a little broken in that it does not pick a correct winner if two players have the same type hand (straight, flush, pair).