Version 2 of the Poker game
It currently has all of the flags to tell the user who one. Logic is for the most part setup well. The problem is that if the winning player has a pair, flush, ect. and another player also has the same set (but different cards). The program will favor and thus declare the player with the higher player number as the winner.